test: move and clean up tests
